requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
boto3==1.40.31
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
tenacity==9.1.2
json-repair==0.50.1
langchain-core==0.3.76
langchain-chroma==0.2.5
langchain-huggingface==0.3.1
chromadb==1.0.20
unstructured[pdf]==0.18.14
sentence-transformers==3.0.1
pillow
pytesseract